+++ src/main/java/act/app/App.java
@@ -800,7 +800,7 @@ public class App extends LogSupportedDestroyableBase {
                 if (!isDevColdStart()) {
                     runnable1.run();
                 } else {
-                    jobManager.now("post_di_load_init", runnable1, true);
+                    jobManager.now("post_di_load_init", runnable1);
                 }
                 Runnable runnable2 = new Runnable() {
                     @Override