+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # ActFramework Change Log
 
 **1.8.18**
+* Fix JobContext init logic
+* Error enhancing bytecode at com.sun.mail.util.logging.MailHandler::publish() #1048
 * ContextualJob cause JobContext get cleared twice #1046
 * JobContext lost when invoking another job #1044
 * Send out Login and Logout event #1043